refactor: split the rendering and save classes to wp-decker's granularity
The four-class split closed every alert on Documentate_Documents but left two
new ones behind: Meta_Boxes at complexity 204 and Meta_Saver at 138, against a
threshold of 100. Running PHPMD over wp-decker's own custom-post-types with
this project's ruleset settles what "like wp-decker" means - it reports zero
alerts, and its largest class there is Decker_Tasks at 48. The reference keeps
classes an order of magnitude smaller than what the first split produced.

Four more extractions, each a group that reached nothing outside itself and
touched no properties, so all four are static helpers:

  Documentate_Document_Field_Help       36  leading text, description,
                                            validation message and the ids
                                            tying them to a control
  Documentate_Document_Scalar_Field     35  text, number, date, select,
                                            checkbox, textarea, TinyMCE
  Documentate_Document_Repeater_Field   45  rows, the clone template, and one
                                            control per column
  Documentate_Document_Content_Writer   95  composing post_content, and the
                                            sanitizing the save path shares

Order mattered. The repeater closure was 33 methods while the help subsystem
was still inline and 12 once it was out, so each boundary was recomputed after
the previous move rather than planned up front.

includes/custom-post-types/ now reports no PHPMD alerts at all, from four when
this branch started. Plugin-wide 52 to 48.

Two things worth saying plainly. Content_Writer at 95 clears the threshold but
is still twice wp-decker's largest; the sanitizers travel with it because both
composing and saving call them, and separating those needs a decision about
where sanitizing belongs rather than another mechanical move.
Documentate_Document_Repeater_Field also exposes prepare_field_control() and
the title helpers, because both row builders share that resolution - it reads
oddly on a class named after repeaters, and the docblock says so.

filter_post_data_compose_content() became private when its group moved, which
would have broken the CPT delegator; the entry point is public and the
delegator now points at the writer.

Verified as before: the editor markup, a composed post_content and the admin
list columns and views are identical to the pre-split code.
