+++ tests/unit/includes/custom-post-types/DocumentateDocumentsHelpersTest.php
@@ -61,7 +61,7 @@ class DocumentateDocumentsHelpersTest extends WP_UnitTestCase {
 	 * @return object
 	 */
 	private function owner_of( $name ) {
-		foreach ( array( $this->documents, $this->meta_boxes, $this->meta_saver, $this->admin_list ) as $candidate ) {
+		foreach ( array( $this->documents, $this->meta_boxes, $this->meta_saver, $this->admin_list, 'Documentate_Document_Content_Writer', 'Documentate_Document_Field_Help', 'Documentate_Document_Repeater_Field', 'Documentate_Document_Scalar_Field' ) as $candidate ) {
 			if ( method_exists( $candidate, $name ) ) {
 				return $candidate;
 			}
@@ -79,10 +79,10 @@ class DocumentateDocumentsHelpersTest extends WP_UnitTestCase {
 	 */
 	private function invoke( $name, array $args = array() ) {
 		$target = $this->owner_of( $name );
-		$method = ( new ReflectionClass( $target ) )->getMethod( $name );
+		$method = new ReflectionMethod( $target, $name );
 		$method->setAccessible( true );
 
-		return $method->invokeArgs( $target, $args );
+		return $method->invokeArgs( $method->isStatic() ? null : $target, $args );
 	}
 
 	/**
